refactor: 가격 0으로 나온 문제 해결,  검색Cell 가격Label 상승 + 하락 - 문자 표시 추가  #1
- string to Int 중 소수점도있는 string이 있어 0 으로 나오는 문제 발생 string to Double로 문제를 해결

+++ Coin/Extension/String+Extension.swift
@@ -9,7 +9,7 @@ import Foundation
 
 extension String {
     func convertPriceKRW() -> String {
-        let convertNSNumber = NSNumber(value: Int(self) ?? 0)
+        let convertNSNumber = NSNumber(value: Double(self) ?? 0.0)
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
         numberFormatter.maximumFractionDigits = 1
